refactor: tableFormat
