+++ frontend/src/lib/tableFormat/cellStyles.scss
