+++ frontend/src/lib/tableFormat/formatAuthorship.tsx
