+++ frontend/src/lib/tableFormat/formatObjectTitle.jsx
@@ -1,8 +1,7 @@
+import { Spacer } from '@components';
 import styled from 'styled-components';
 
 import nebula from '/src/nebula';
-
-import { Spacer } from '../components';
 
 const TitleNote = styled.span`
   color: var(--color-text-dim);
