+++ frontend/src/lib/tableFormat/formatRundownSymbol.jsx
