+++ frontend/src/pages/Rundown/RundownTable.jsx
@@ -1,15 +1,14 @@
+import nebula from '/src/nebula';
+import { Table } from '@components';
+import { useDialog } from '@features/Dialogs';
+import { formatRowHighlightColor, formatRowHighlightStyle } from '@lib/tableFormat';
 import { useMemo, useRef, useEffect } from 'react';
 import { useSearchParams, useLocation } from 'react-router-dom';
 
-import nebula from '/src/nebula';
-import { Table } from '@components';
-import { formatRowHighlightColor, formatRowHighlightStyle } from '/src/tableFormat';
-
-import { useDialog } from '@features/Dialogs';
-import { useNebula } from '@/features/Nebula';
-
 import RundownTableWrapper from './RundownTableWrapper';
 import { getRunModeOptions, getRundownColumns } from './utils';
+
+import { useNebula } from '@/features/Nebula';
 
 const RundownTable = ({
   data,
